+++ server/src/server.mjs
@@ -1,38 +1,40 @@
-import dotenv from "dotenv";
+// import dotenv from "dotenv";
 import express from "express";
 import mime from "mime-types";
-import twilio from "twilio";
+// import twilio from "twilio";
 
 // import affirmationsRouter from "./affirmationsRouter.mjs";
 import jwtCheck from "./jwtCheck.mjs";
 // 2nd API
 import quoteRouter from "./quoteRouter.mjs";
+import smsRouter from "./smsRouter.mjs";
 import taskRouter from "./taskRouter.mjs";
 import userRouter from "./userRouter.mjs";
 
 const app = express();
 
 // Twilio
-dotenv.config();
+// dotenv.config();
 
-const accountSid = process.env.TWILIO_ACCOUNT_SID;
-const authToken = process.env.TWILIO_AUTH_TOKEN;
-let client = twilio(accountSid, authToken);
+// const accountSid = process.env.TWILIO_ACCOUNT_SID;
+// const authToken = process.env.TWILIO_AUTH_TOKEN;
+// let client = twilio(accountSid, authToken);
 
 // const phone = await db.getPhone(request.user.sub);
 // response.json(tasks);
 
 // comment
-client.messages.create({
-  body: "Howdy!",
-  from: "+12178852760",
-  to: "+12246027354",
-});
+// client.messages.create({
+//   body: "Howdy!",
+//   from: "+12178852760",
+//   to: "+12246027354",
+// });
 // .then(message => console.log(message.sid));
 
 // app.use("/api/affirmations", jwtCheck, affirmationsRouter);
 app.use("/api/tasks", jwtCheck, taskRouter);
 app.use("/api/users", jwtCheck, userRouter);
+app.use("/api/sms", jwtCheck, smsRouter);
 // 2nd API
 app.use("/api/quote", jwtCheck, quoteRouter);
